--- _examples/adsb/mock_iq/go.mod
@@ -1,21 +1,0 @@
-module github.com/ml4wireless/specpipe/_examples/mock_iq
-
-go 1.20
-
-require (
-	github.com/ThreeDotsLabs/watermill v1.2.0
-	github.com/ThreeDotsLabs/watermill-nats/v2 v2.0.2
-	github.com/nats-io/nats.go v1.31.0
-)
-
-require (
-	github.com/google/uuid v1.3.1 // indirect
-	github.com/klauspost/compress v1.17.1 // indirect
-	github.com/lithammer/shortuuid/v3 v3.0.7 // indirect
-	github.com/nats-io/nkeys v0.4.6 // indirect
-	github.com/nats-io/nuid v1.0.1 // indirect
-	github.com/oklog/ulid v1.3.1 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	golang.org/x/crypto v0.14.0 // indirect
-	golang.org/x/sys v0.13.0 // indirect
-)
