+++ _examples/iq_adsb/requirements.txt
@@ -1,0 +1,20 @@
+annotated-types==0.6.0
+anyio==4.3.0
+certifi==2024.2.2
+charset-normalizer==3.3.2
+click==8.1.7
+fastapi==0.110.1
+h11==0.14.0
+idna==3.6
+nats-py==2.7.2
+pydantic==2.6.4
+pydantic_core==2.16.3
+requests==2.31.0
+sniffio==1.3.1
+SpeechRecognition==3.10.3
+starlette==0.37.2
+typing_extensions==4.11.0
+urllib3==2.2.1
+uvicorn==0.29.0
+websockets==12.0
+pytz==2024.1
